Remove unwrap in image type clear

+++ src/ui/images/ui_image_download.rs
@@ -4,9 +4,9 @@ use super::{
     hasimagekey::HasImageKey,
     image_select_state::ImageSelectState,
     pages::{
-        handle_correct_grid_request, handle_grid_change, handle_shortcut_selected,
-        render_page_pick_image, render_page_shortcut_images_overview,
-        render_page_shortcut_select_image_type, render_page_steam_images_overview, handle_image_selected,
+        handle_correct_grid_request, handle_grid_change, handle_image_selected,
+        handle_shortcut_selected, render_page_pick_image, render_page_shortcut_images_overview,
+        render_page_shortcut_select_image_type, render_page_steam_images_overview,
     },
     possible_image::PossibleImage,
     texturestate::TextureDownloadState,
@@ -157,7 +157,7 @@ impl MyEguiApp {
                 self.handle_image_type_selected(image_type);
             }
             UserAction::ImageSelected(image) => {
-                handle_image_selected(self,image);
+                handle_image_selected(self, image);
             }
             UserAction::BackButton => {
                 self.handle_back_button_action();
@@ -228,32 +228,31 @@ impl MyEguiApp {
     }
 
     fn handle_image_type_clear(&mut self, image_type: ImageType) {
+        let app_id = self
+            .image_selected_state
+            .selected_shortcut
+            .as_ref()
+            .map(|s| s.app_id());
         let data_folder = &self
             .image_selected_state
             .steam_user
             .as_ref()
-            .unwrap()
-            .steam_user_data_folder;
-        for ext in POSSIBLE_EXTENSIONS {
-            let file_name = image_type.file_name(
-                self.image_selected_state
-                    .selected_shortcut
-                    .as_ref()
-                    .unwrap()
-                    .app_id(),
-                ext,
-            );
-            let path = Path::new(data_folder)
-                .join("config")
-                .join("grid")
-                .join(file_name);
-            if path.exists() {
-                let _ = std::fs::remove_file(&path);
+            .map(|s| &s.steam_user_data_folder);
+        if let (Some(app_id), Some(data_folder)) = (app_id, data_folder) {
+            for ext in POSSIBLE_EXTENSIONS {
+                let file_name = image_type.file_name(app_id, ext);
+                let path = Path::new(data_folder)
+                    .join("config")
+                    .join("grid")
+                    .join(file_name);
+                if path.exists() {
+                    let _ = std::fs::remove_file(&path);
+                }
+                let key = path.to_string_lossy().to_string();
+                self.image_selected_state.image_handles.remove(&key);
             }
-            let key = path.to_string_lossy().to_string();
-            self.image_selected_state.image_handles.remove(&key);
+            self.image_selected_state.image_type_selected = None;
         }
-        self.image_selected_state.image_type_selected = None;
     }
 
     fn handle_set_game_mode(&mut self, game_mode: GameMode) {
@@ -304,11 +303,6 @@ impl MyEguiApp {
         };
     }
 
-    
-
-    
-
-    
     fn handle_back_button_action(&mut self) {
         let state = &mut self.image_selected_state;
         if state.possible_names.is_some() {
